🐛 Fix when we start appcues from an activity the first activity was not the current
