+++ appcues/build.gradle
@@ -59,6 +59,7 @@ dependencies {
 
     implementation "androidx.appcompat:appcompat:1.4.1"
     implementation 'androidx.activity:activity-compose:1.4.0'
+    implementation "androidx.startup:startup-runtime:1.1.0"
 
     implementation "androidx.compose.ui:ui:1.0.5"
     implementation "androidx.compose.ui:ui-tooling:1.0.5"
