+++ appcues/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.appcues">
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -9,6 +10,17 @@
             android:name=".ui.AppcuesActivity"
             android:exported="false"
             android:theme="@style/Appcues.AppcuesActivityTheme" />
+
+        <provider
+            android:name="androidx.startup.InitializationProvider"
+            android:authorities="${applicationId}.androidx-startup"
+            android:exported="false"
+            tools:node="merge">
+
+            <meta-data
+                android:name="com.appcues.AppcuesInitializer"
+                android:value="androidx.startup" />
+        </provider>
     </application>
 
 </manifest>
