+++ appcues/src/main/java/com/appcues/monitor/CustomerActivityMonitor.kt
@@ -1,14 +1,9 @@
 package com.appcues.monitor
 
-import android.app.Activity
-import android.app.Application
-import android.content.Intent
-import android.os.Bundle
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.savedstate.SavedStateRegistryOwner
 import com.appcues.R
 import com.appcues.di.AppcuesKoinComponent
-import com.appcues.di.getApplicationContext
 import com.appcues.di.getOwnedViewModel
 import com.appcues.domain.entity.Experience
 import com.appcues.domain.gateway.CustomerViewGateway
@@ -19,17 +14,11 @@ import org.koin.androidx.viewmodel.ViewModelOwner
 
 internal class CustomerActivityMonitor(
     override val scopeId: String,
-) : CustomerViewGateway, ActivityMonitor, Application.ActivityLifecycleCallbacks, AppcuesKoinComponent {
-
-    init {
-        getApplicationContext().registerActivityLifecycleCallbacks(this)
-    }
-
-    private var customerActivity: Activity? = null
+) : CustomerViewGateway, ActivityMonitor, AppcuesKoinComponent {
 
     override suspend fun showExperience(experience: Experience) {
         withContext(Dispatchers.Main) {
-            customerActivity?.let {
+            AppcuesActivityMonitor.activity?.let {
                 it.startActivity(AppcuesActivity.getIntent(it, scopeId, experience))
                 it.overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
             }
@@ -37,7 +26,7 @@ internal class CustomerActivityMonitor(
     }
 
     override fun getCustomerViewModel(): CustomerViewModel? {
-        return customerActivity?.let {
+        return AppcuesActivityMonitor.activity?.let {
             getOwnedViewModel(
                 owner = {
                     ViewModelOwner.from(
@@ -47,35 +36,5 @@ internal class CustomerActivityMonitor(
                 }
             )
         }
-    }
-
-    override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-        if (activity !is AppcuesActivity) {
-            this.customerActivity = activity
-        }
-    }
-
-    override fun onActivityStarted(activity: Activity) {
-        // do nothing
-    }
-
-    override fun onActivityResumed(activity: Activity) {
-        // do nothing
-    }
-
-    override fun onActivityPaused(activity: Activity) {
-        // do nothing
-    }
-
-    override fun onActivityStopped(activity: Activity) {
-        // do nothing
-    }
-
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
-        // do nothing
-    }
-
-    override fun onActivityDestroyed(activity: Activity) {
-        activity.sendLocalBroadcast(Intent(activity.intentActionFinish()))
     }
 }
